Refactor: 카트 위치 및 카메라 동기화 로직 useSyncCartAndCamera 훅으로 분리
